+++ src/components/Pagination.js
@@ -1,4 +1,4 @@
-import { Pagination } from 'twreporter-react-components'
+import { Pagination } from '@twreporter/react-components'
 import PropTypes from 'prop-types'
 import React, { PureComponent } from 'react'
 import get from 'lodash/get'