+++ src/constants/category-strings.js
@@ -5,6 +5,6 @@ export default {
   livingAndMedicalCare: '生活．醫療',
   cultureAndArt: '文化．藝術',
   international: '國際．兩岸',
-  reviews: '觀點',
+  reviews: '評論',
   infographic: '多媒體'
 }