+++ CHANGELOG.md
@@ -1,9 +1,11 @@
 ### Unreleased
 - Fix Google structure data errors
 
-### 2.3.0-beta.1.0.0
-- Create registration system
+### 3.0.0
+- Create registration system (RS)
+- RS: SignIn, SignUp, Activation...
 - Implement bookmark feature
+- Mobile Navbar
 
 ### 2.2.1
 - Update the schema of homepage [#665](https://github.com/twreporter/twreporter-react/pull/665)